normalized camstats, added functional tests
