+++ isis/src/base/objs/CameraStatistics/CameraStatistics.cpp
@@ -441,6 +441,7 @@ namespace Isis {
 
     PvlGroup pAspectRatio("AspectRatio");
     pAspectRatio += constructKeyword("AspectRatioMinimum", m_aspectRatioStat->Minimum());
+    // Note: Maximum is spelled wrong here.
     pAspectRatio += constructKeyword("AspectRatioMaximun", m_aspectRatioStat->Maximum());
     pAspectRatio += constructKeyword("AspectRatioAverage", m_aspectRatioStat->Average());
     pAspectRatio += constructKeyword("AspectRatioStandardDeviation",
@@ -513,4 +514,3 @@ namespace Isis {
     return returnPvl;
   }
 }
-
